// src/gql/graphql.ts
export type Stage =
  /** The Draft is the default stage for all your content. */
  | 'DRAFT'
  /** The Published stage is where you can publish your content to. */
  | 'PUBLISHED';
// ...
export type ProductItemFragment = { id: string, name: string, description: string, price: number, averageRating: number, categories: Array<{ name: string }>, images: Array<{ url: string }>, reviews: Array<{ id: string, name: string, headline: string, email: string, content: string, rating: number, stage: Stage }> };
// ...
export type ReviewsItemFragment = { id: string, name: string, headline: string, email: string, content: string, rating: number, stage: Stage };